requirements.txt
aiohttp>=3.8
discord.py>=2.1.0